DeviceInfo: Target our own build number
Tracking the build number is pointless on a non-nexus device, and even
on a Nexus device (The new ones atleast) the vendor version is the build
number, which leaves you with duplicate information. Users can also look
at the security patch level to determine the Android build. There's just
too many things that inform you of the same thing.

So let's make our own build version system, that ties into our current
versions. We have the CypherOS version but nothing specifying what build
the user is running. The build date could  be a factor, but now all
devices get compiled on the same day so that's a setback. I'm not 100%
sure how i want this build system to work, but this is a start.

Signed-off-by: Chris <sirhc95@cypheros.co>